Runed Blog Migration, added: form of creating posts, retrievement of posts, ability to add image

+++ database/migrations/2020_01_23_162946_create_blogs_table.php
@@ -15,7 +15,7 @@ class CreateBlogsTable extends Migration
     {
         Schema::create('blogs', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->unsignedInteger('author');
+            $table->unsignedBigInteger('author');
             $table->mediumText('title');
             $table->longText('body');
             $table->string('image')->nullable();
